+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## UNRELEASED
+
+- Abstracted `extend` into the compose-extend package.
+
 ## 1.0.1
 
 - Deprecate docEvents. Not that useful. Rather confusing. Much complexity.